pyproject.toml
[tool.poetry]
name = "netnet"
version = "0.1.0"
description = ""
authors = ["stefanlight <steffmukhin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quant = {git = "https://github.com/MagM1go/quant.git"}
aiohttp = "^3.8.6"
python-dotenv = "^1.0.0"
uvloop = "^0.19.0"
typing-extensions = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
